Unify generate name functions to be used in the frontend
