+++ src/hooks/use-site-details.tsx
@@ -15,6 +15,7 @@ import { useContentTabs } from 'src/hooks/use-content-tabs';
 import { useIpcListener } from 'src/hooks/use-ipc-listener';
 import { useOffline } from 'src/hooks/use-offline';
 import { simplifyErrorForDisplay } from 'src/lib/error-formatting';
+import { generateNumberedName } from 'src/lib/generate-site-name';
 import { getIpcApi } from 'src/lib/get-ipc-api';
 import { useAppDispatch } from 'src/stores';
 import { snapshotThunks } from 'src/stores/snapshot-slice';
@@ -483,7 +484,10 @@ export function SiteDetailsProvider( { children }: SiteDetailsProviderProps ) {
 				);
 			};
 
-			const finalSiteName = await getIpcApi().generateCopySiteName( sourceSiteId );
+			const finalSiteName = await generateNumberedName(
+				sprintf( __( '%s Copy' ), sourceSite.name ),
+				sites
+			);
 
 			const tempSiteId = crypto.randomUUID();
